+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 ### Added
 
 - new --assignee flag on `issue list` allowing you to list issues assigned to
+- `issue delete` command to delete issues by id
+
   @me, @all or a given user. A flag is shorthand for --assignee @all
 
 ## [0.6.4] - 2025-08-12